refactor(toolkit-lib,cli): access default messages through IoHelper (#513)
This change is in preparation of work to have all default message go
through `IoHelper` instead of using `IO.DEFAULT_XYZ`. It contains a
number of clean-up and alignment changes:

- `IoDefaultMessages` is now a local-only class, that is only accessed
through the new `IoHelper.defaults` getter.
- Removed unused usage of `IoHelper` and `IoDefaultMessages` in
`DiffFormatter`.
- Changed every use of `IoDefaultMessages` to `IoHelper.defaults`
- Made `IoDefaultMessages` methods async to be consistent with the rest,
we can `void` at the place of use if we really need to
- Also fixed most of the remaining `void`'ed call-sites for the above.
Only the legacy logging is now left.
- Removed unused `messages.ts` file
- Removed unused `trace` from `logging.ts` (and respective test) to not
encourage further use

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license