fix(toolkit-lib): deploy options is missing `importExistingResources` and `hotswapProperties.ecs.stabilizationTimeoutSeconds` (#520)
We had two version of the `DeploymentMethod` interface and recent
features inadvertently only added to one interface. This meant two new
features were previously not available in `toolkit-lib`.

This PR fixes the issue and prevents future misses by unifying on the
same interface. Also fixes a bug with `hotswapProperties.ecs` not being
optional and aligning some code.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license