feat(toolkit-lib)!: hotswap is now a deployment method (#517)
Historically we modeled hotswap deployments in the CLI as a boolean flag
on the deploy command. At some point we then added control over the
procedure for not hotswappable resources: fallback to a regular CFN
deployment or do nothing. Most of this design was carried over to
`toolkit-lib`.

This change re-models hotswap deployments as a deployment method on the
same level as change-set and direct deployments. This make sense, since
hotswap also takes additional input that is otherwise ignored. For its
fallback mode, the hotswap deployment method can take any of the other
deployment methods (but not itself).

BREAKING CHANGE: hotswap deployments are now modeled as a deployment
method. Instead of `DeployOptions.hotswap`, use
`DeployOptions.deploymentMethod = { method: 'hotswap' }`. Similarly,
`DeployOptions.hotswapProperties` is now `DeployOptions.deploymentMethod
= { method: 'hotswap', properties: myHotswapProperties }`.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license