fix: property references are skipped when `DependsOn` is present (refactor) (#524)
When computing the graph for the stack, if there is a `DependsOn`, the
algorithm will ignore other dependencies. Also, `DependsOn` may be a
string or an array, and this was not being properly handled.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license