feat(toolkit-lib)!: generic messages do not have a code (#526)
In the [original
RFC](https://github.com/aws/aws-cdk-rfcs/blob/main/text/0300-programmatic-toolkit.md)
we had introduced the concept of "standard" messages for which no
backwards compatibility rules won't apply. In the first developer
preview release of `toolkit-lib` this was implemented as message with a
code ending in `0000` being consider "standard" (or "default") and other
messages.

With the experience from the developer preview, we have found this way
to handle the distinction slightly confusing and putting comparatively
complicated work on the user (they need to know and check for the code
ending in `0000`). It was also easy to accidentally make a mistake by
copy and pasting the message code and not noticing it's specialness.

Instead, we are now proposing to change this implementation: Messages
and requests with a code will be subject to our backwards compatibility
guarantees and can be relied upon by integrators. Standard messages will
no longer have a code and must be treated purely informational. Note
that all requests and any messages with a payload will always have a
code. However some coded messages might not currently have a payload
(but may in future according to our backwards compatibility rules).

This PR reflects these changes and updates the message registry with the
respective documentation.

BREAKING CHANGE: The `code` field on messages is now optional. All
messages with a code are subject to our backwards compatibility
guarantees and can be relied upon by integrators. Messages that
previously had a "default code" ending in `0000` no longer have a code
and must be treated purely informational. Requests and messages with a
payload are always coded.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license