feat: digest computation for refactor considers cross-stack references (#529)
Previously the digest computation was being done stack by stack. One of
the implications of this is that cross-stack references were ignored.
Specifically, `Fn::ImportValue` was treated as a regular value, rather
than as a reference.

Change that, so we consider all stacks at once, and treating cross-stack
references in the same way as within-stack references are treated, for
the purposes of computing the digests of all resources.

While we're at it, the purely graph related code was extracted to its
own class, `ResourceGraph`.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license