refactor(cli): create `ProxyAgent` in the CLI in preparation of `proxy-agent` removal from toolkit-lib (#532)
Updates the `CdkToolkit` and `Notices` APIs to take a Node https.Agent
instead of options. This gives more flexibility to integrators while
retaining proxy support. We can make the changes to `Notices` since it
is not a public API.

The CLI changed need to be split out from the toolkit-lib changes, so
the latter can be safely marked as breaking without touching CLI files.
The respective `toolkit-lib` changes are in #533

Relates to #398 

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license