fix(integ-testing): sporadic timeout of `test resource import with construct that requires bundling` (#536)
Increase the timeout of this specific test to 30min.

-----------------------

We are seeing occasional timeouts of this test in our canary runs, with
this being the cause of the last 4 canary failures.

```console
Summary of all failing tests
--
2062 | FAIL ./cdk-test-resource-import-with-construct-that-requires-bundling.integtest.js (1215.923 s)
2063 | ● test resource import with construct that requires bundling
2064 |  
2065 | Timeout: test took more than 1200s to complete
2066 |  
2067 | at Timeout._onTimeout (../../lib/with-timeout.ts:23:18)
```

Looking at the duration of the test in successful runs:

- `[INTEG TEST::test resource import with construct that requires
bundling] Done (18m48s).`
- `[INTEG TEST::test resource import with construct that requires
bundling] Done (19m45s)`
- `[INTEG TEST::test resource import with construct that requires
bundling] Done (18m59s)`

The prolonged durations stem from bundling in the `NodejsFunction`
construct (only used by this test):

```console
#18 naming to docker.io/library/cdk-8a998ad57638856c93a312f4354d3295aec983047a6cb011b5dc22b6f24e8dc6 done
--
1357 | #18 DONE 1.2s
1358 | Bundling asset cdktest-0gw7l6udpnlr-importable-stack/cdk-import-nodejs-lambda-test/Code/Stage...
1359 |  
1360 | asset-output/index.js  169b
1361 |  
1362 | ⚡ Done in 3ms
1363 |  
1364 | ✨  Synthesis time: 829.86s
```

Specifically, the extraction of the image seems to be taking long:

```console
...
...
#5 extracting sha256:aaa6e8fa0ae46c5f73ab02a77b4e825c21c375e4848f072cb1f28787d3ac5c71 307.6s
--
1278 | #5 extracting sha256:aaa6e8fa0ae46c5f73ab02a77b4e825c21c375e4848f072cb1f28787d3ac5c71 309.9s done
```

Given a 20min timeout, this is probably a little too close for comfort.
Lets extend the timeout a bit so that this test can run smoothly on
smaller CodeBuild machines.

> Note: We don't want to increase the machine size because canary runs
don't have to be fast - so lets save some money.

## Note

In PR builds, where we run on GitHub runners, and use a bigger machine,
we see much faster durations:

- pass ✅ | test resource import with construct that requires bundling |
8m30s



---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license