feat(toolkit-lib)!: support custom `https.Agent` for SDK requests (#533)
Fixes #398 
Depends on #532

Previously automatically configured the SDK to pass all requests through
an instance of `ProxyAgent` from the third-party `proxy-agent` package.
A user could configure its proxy settings, but by default it would act
as a pass through. `proxy-agent` is a hefty dependency (primarily due to
its support for `pac` files) that not everyone might want to use. This
is the standard interface in Node.js and we are aligning on it.

Note that with this we are also removing built-in (but not advertised)
support for managing proxy configuration through env vars like
`HTTP_PROXY` or `HTTPS_PROXY`. This functionality was provided by
`proxy-agent` and the easiest way to get it back is by configuring
`proxy-agent` again.

BREAKING CHANGE: Dependency on `proxy-agent` has been removed, this also
removes the support of configuring a proxy through like `HTTPS_PROXY`
and `AWS_CA_BUNDLE`. The type of `sdkConfig.httpOptions` has changed to
now optionally take a single `agent: https.Agent` property instead of
the proxy settings. The restore the previous behavior, pass an instance
of `ProxyAgent` to `sdkConfig.httpOptions`: `{ agent: new ProxyAgent({
ca: oldHttpOptions.caBundlePath, getProxyForUrl: () =>
oldHttpOptions.proxyAddress }) }`

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license