feat(toolkit-lib)!: improved types for BaseCredentials (#542)
This PR refactors the `BaseCredentials` implementation to provide better
type safety and a cleaner API surface. By converting from an abstract
class to a static factory pattern with proper interfaces, we improve the
developer experience and make the code more maintainable.

The changes also reduce external dependencies by removing the direct
dependency on `@smithy/node-http-handler` and creating our own
simplified interface for request handler settings.

Reason for this change is that in the previous version
`SdkProviderServices` was used but not publicly exported since this
interface wasn't meant for public consumption. That however made it
difficult to implement custom `BaseCredentials` solutions. Instead of
`SdkProviderServices` we have no reduced the public API surface to the
exactly needed properties.

BREAKING CHANGE: Custom implementations of `BaseCredentials` have
changed. If you previously extended the abstract `BaseCredentials` class
and implemented the abstract method, you can now pass any object
implementing the new `IBaseCredentialsProvider` interface. The method
name change to `sdkBaseConfig` and it now receives two parameters: an
`ioHost` that can be used to emit messages and a `clientConfig` to
should be honored for all SDK requests. However the `clientConfig` may
be extended as needed.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license