chore(toolkit-lib): replace typedoc with api-extractor (#543)
Removes typedoc configuration in favor of api-extractor. We were running
both in parallel for a transition period, but now only api-extractor is
used. Also includes related updates to the toolkit-lib package
configuration.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license