feat(toolkit-lib): emit drift in span (#550)
When we merged #442 we forgot to add message spans to the drift command.
This PR rectifies this.

Also contains a refactor of `MessageSpan` to be a separate class and to
represent itself as `IoHelper`. With that we can now pass a `Span` down
into subroutines and have those messages also emitted as part of the
span. Start using that in some places.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license