fix(toolkit-lib): stack definitions for refactor are not considering resource refrences (#547)
The `createStackRefactor` method, from the CloudFormation API, requires
the client to pass not only the mappings, but also the templates in the
state they are supposed to be after the refactor.

Previously, this was done by iterating over the mappings (either
computed by the toolkit or prescribed by the user), and using them as a
list of instructions on how to modify the deployed stacks, moving
resources around as necessary. This is a simple way to put the resources
in their correct locations. But it becomes unwieldy when it comes to
updating references between resources. Every time a resource is moved,
all references to it -- and sometimes from it -- have to be updated as
well.

This PR changes implements a different algorithm: take all the
synthesized templates as a starting point, and adjust them so that the
final templates are equal to the deployed ones up to logical IDs and
references. The result is what will be sent to CloudFormation.
Specifically this involves:

- Removing resources that exist locally but not remotely.
- Adding resources that exist remotely but not locally.
- Update the CDK construct path to the deployed value. (we will drop
this once CloudFormation starts allowing changes in the construct path
along with refactors).

One case that is not yet covered is when physical IDs are set. Because
we use the physical ID plus the type of the resource to identify it,
users are free to change other properties, and we can still recognize
that it's the same resource. Since we are simply taking the local
resource as the basis and not updating anything other than the construct
path, it's possible that it differs from the deployed one. This case
will be handled in a following PR.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license