chore: stack refactor code re-organization (#556)
This is a major re-organization of the stack refactoring code. The core
change was the introduction of the `RefactoringContext`, that
encapsulates the data and operations needed for a refactor, for a single
environment. In future changes, this class will also start exposing
methods to execute the refactor.

No behavior was changed, except in the presentation layer: before, all
mappings were grouped in a single table. Now, there is a section for
each environment, with its own table, to make it clearer to the user
that each one is a different refactor operation:

![Screenshot 2025-05-30 at 11 02
25](https://github.com/user-attachments/assets/d4ac65e6-7d7c-4e3e-9438-a5973b27e572)


---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license