fix(toolkit-lib): `refactor` feature is not marked as unstable (#563)
This introduces the concept of unstable featues to the toolkit. The CLI
always opts into it because it has its own similar flag.

Over time, the type `UnstableFeature` will change according to the
features that are unstable at the moment. All commands that are
considered unstable should call the `requireUnstableFeature` method.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license