fix(toolkit-lib): new `mappingSource` option to replace multiple mutual exclusive options in the refactor API (#559)
Introduce a new type, `MappingSource`, that allows the user to tell how
they want the toolkit to get the mappings: computing it, using an
explicitly provided one, or using the reverse of an explicitly provided
one.

This replaces three previous separate properties: `exclude`, `mappings`
and `revert` with a single `mappingSource`.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>