chore: fix typo (#566)
There should be a linter rule against this but I don't have time to
write it. Without the extra `*` the documentation does not become part
of the docstring for the property.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license