chore(cli): type `currentAction` as `CliAction` (#571)
`CliAction` type feels slightly off as it should be `CliAction` and not
`ToolkitAction`. This makes the default make sense. Either way, the
internal property of `CliIoHost` is typed to `CliAction`.
 
In practice this probably doesn't matter because I don't think we are
initializing `CliIoHost` with anything but `none`.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Co-authored-by: Momo Kornher <kornherm@amazon.co.uk>