chore: improved generation of stack definitions (#584)
The Cloudformation refactoring API needs, in addition to the mappings,
the
resulting templates for each affected stack. The resulting templates are
basically the synthesis produced, but with some differences:

- Resources that exist in the local stacks, but not in the remote
stacks, are
  not included.
- Resources that exist in the remote stacks, but not in the local
stacks, are
  preserved.
- For resources that exist in both stacks, but have different
properties, the
deployed properties are used, but the references may need to be updated,
if
  the resources they reference were moved in the refactoring.

Why does the last difference exist, to begin with? By default, to
establish
whether two given resources are the same, roughly speaking we compute
the hash
of their properties and compare them. But there is a better source of
resource
identity, that we can exploit when it is present: the physical name. In
such
cases, we can track a resource move even if the properties are
different, as
long as the physical name is the same.

The process of computing the resulting templates consists in:

1. Computing a graph of deployed resources.
2. Mapping edges and nodes according to the mappings (that we either 
   computed or got directly from the user).
3. Computing the resulting templates by traversing the graph and
   collecting the resources that are not mapped out, and updating the
   references to the resources that were moved.




---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license