chore(integ-testing): test CDK on multiple versions of Node (#591)
We weren't running our integ tests on all Node versions. Add that.

Unfortunately, the success or failure of the integ test run on this PR
means nothing, because the workflow only takes effect after merging to
`main`.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license