chore(integ-runner): prepare integ-runner for different engines (#605)
Updates the engine contract `ICli` to support other engines. The main
change here is to make all methods async. For `watch` this required some
more changes and a switch to a callback based system instead of
returning and EventEmitter. Finally the PR changes the return type of
`list` to be an array of stack names instead of a newline separated
string.

Due to the "viral" nature of async/await, this touches a lot of code
paths.

Contains some related clean-up:

- simplified the mock engine `MockCdkMocks` to not depend on the anyway
unused `CdkCliWrapper` any more
- set `isolatedModules` in tsconfig instead of jest in `integ-runner`
- Remove duplicated command input types from `cdk-cli-wrapper` in favor
of the same types from `@aws-cdk/cloud-assembly-schema`

After this change is in, it will be possible to implement an engine
based on toolkit-lib.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license