feat(toolkit-lib): `fromAssemblyDirectory` can ignore missing context (#611)
Previously `fromAssemblyDirectory` created a
`ContextAwareCloudAssemblySource` with lookups disabled. This would
always fail if there was any missing context in the assembly, because
you can't perform lookups with just the CloudAssembly directory.

This PR updates the logic to use a `CachedCloudAssembly` instead and
check for missing context. The missing context check can be disabled
using the new `failOnMissingContext` option (defaults to fail).

Resolves #549

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Co-authored-by: Momo Kornher <mail@moritzkornher.de>
Co-authored-by: Momo Kornher <kornherm@amazon.co.uk>