feat(integ-runner): support a new toolkit-lib based engine that does not require CDK CLI to be installed (#644)
The new engine will eventually unlock performance optimizations and in
generally be more consistent with CLI behavior. Use like this:

```console
integ-runner --unstable toolkit-lib-engine
```

For now, this change focuses on feature parity as there are some
optimizations that will be tricky to do without refactoring the codebase
and/or having a bunch of conditionals in the code base.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license