fix(cli): logs entire environment (#623)
Right now the CLI logs the entire environment to debug logging. This is
a lot of data, and potentially also includes tokens we'd rather don't
end up in logging.

Instead, just log our additions to the existing environment.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license