chore: remove node-bundle in favor of node-backpack (#653)
In https://github.com/cdklabs/node-backpack/pull/1167, we moved the code
of this private tool to the external `node-backpack` package.

Remove it here and use the external tool wherever we used to use the
private package.

Closes https://github.com/aws/aws-cdk-cli/issues/17

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license