fix(integ-testing): use specificied library version in init tests (#654)
The init tests used always to use the library version that was baked
into the CLI. The CLI has since grown a `--lib-version` flag that allow
specifying the library version.

Respect the library-under-test version that is being given to the integ
tests, and use that version to test with.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license