fix(cdk-assets): fails installation when new dependencies are released (#603)
Because `cdk-assets` CLI depends on floating versions of dependencies,
it may break when new 3rd party dependencies are released, or when it is
installed just around the time a new version of AWS SDKv3 gets released
(because that has hundreds of packages that all refer to each other by
specific version, but they're subject to publishing/replication latency
so they don't all become instantaneously available).

We used to generate a shrinkwrap file for the `cdk-assets` CLI but
apparently that didn't work because we're still getting reports of
breakage.

Instead, we'll resort to bundling the dependencies of `cdk-assets` into
the CLI itself, so it cannot go wrong.

Because the `cdk-assets` package does double duty as a CLI package as
well as a library package, we can't just naively stick the `BundleCli`
job onto the package: doing that would remove all the bundled
dependencies from the `package.json` list, which breaks its function as
a library.

Instead, we're splitting the `cdk-assets` package in two:

* `cdk-assets`: will continue to contains the CLI, which will be bundled
* `@aws-cdk/cdk-assets-lib`: the core portion of the assets publishing,
available as a library.

Because `cdk-assets` can no longer be used as a library, we have bumped
the major version to `4` to indicate this change in API surface. In CDK
pipelines we install `cdk-assets@latest`, so it will automatically pick
up major version 4 once released.

(While this PR is strictly speaking a breaking change, I'm not adding
any "breaking change" markers into the PR description, because this is
changing many files across multiple packages, and I don't trust the
automatic major version bumping script to not accidentally MV bump
something I didn't intend to)

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>