fix: longer timeout for notices in the testing environment (#656)
We're observing a lot of these errors in our integration tests:

```
[09:29:38] Could not refresh notices: _ToolkitError: Network error: Request timed out
```

It could be that our timeout is too short on a heavily congested
machine. Increase the timeout specifically for those cases to improve
the reliability of our tests.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license