fix: CCAPI provider does not do pagination (#678)
The CCAPI provider stops after the first page of results. This may cause
for example EC2 Prefix Lists that exist to not be found, if they don't
occur in the first page of results.

Make the provider retrieve all pages of results.

Also in this PR:

- New SDK mocks had been added without them being added to all the
places where they needed to be added to be reset properly. Instead, put
them all into an object so we can do a reliable `for` loop that will
never go out of date again.
- Up the `MaxResults` to 100, so we need to do fewer paginations and the
chances of the error message finding too many elements has a reasonable
indicator of the number of resources actually found.

Fixes https://github.com/aws/aws-cdk-cli/issues/587

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license