chore(cli): telemetry sink (#585)
Introduces a generic telemetry sink to the cli. it is tested but not
used anywhere.

There are 3 base types of telemetry sinks:

- `FileSink` -> gathers events and writes them to a file 
- `IoHostSink` -> gathers events and sends them to the IoHost (for
writing to stdout/err)
- `EndpointSink` -> gathers events and sends them to an external
endpoint, batching at intervals of 30 seconds.



---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license