chore(cli): cli type registry json (#682)
Pulling out of #631.

This is relevant for telemetry but also helpful in general to compile
the [dynamic] source of truth in `cli-config.ts` into a static json for
our viewing pleasure.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license