fix(integ-runner): snapshots involving AZs are different if `enableLookups: false` (#685)
In https://github.com/aws/aws-cdk-cli/pull/666 we used `synthFast()` to
always generate a snapshot in the same was as was used for validating
(which was also using `synthFast()`).

The difference being, that the context we load depends on the
`enableLookups: false|true` flag that's passed to `new IntegTest()` in
the test case itself. So when writing the snapshot we have to take that
same field into account.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license