chore: `cdk-assets` did not release (#687)
When we last bumped the cloud assembly schema version, `cdk-assets` did
not release, because it doesn't have a dependency on the cloud assembly
schema package we could track in projen.

Relates to https://github.com/aws/aws-cdk/issues/34912.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>