chore: document warning against trying to be clever (#688)
I was trying to be clever, and that caused a bad experience for
customers.

The reason was that I had forgotten we were jsonschema validating the
manifest by consumers, and so my neat trick wouldn't work.

Document that here so that I won't try again, nor anyone else.


---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>