fix(toolkit-lib): make default environment resolution optional (#689)
This PR makes the default environment resolution optional in the
toolkit-lib. It allows avoiding unnecessary STS calls when the
environment is explicitly specified or when local actions are performed
without internet access.

Then uses the new feature in `integ-runner` to disable the default env
lookup in the toolkit-lib engine. The existing engine doesn't provide
the env variables either, so it's clearly not needed. We don't want to
add to this contract at this time and also resolving the account adds an
unnecessary delay.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

Co-authored-by: Rico Hermans <rix0rrr@gmail.com>