fix(integ-runner): fail with correct error when integ manifest exists but cannot be loaded (#692)
Previously if an integ manifest exists, but could not be loaded for some
reason we would be falling back to the deprecated legacy style tests.
This code path will than always fail with a misleading error:
`"integ-runner" can only operate on apps with a single stack.` The error
is because the modern integ tests always have include stacks, but legacy
tests only handle a single stack or a pragma list.

This change improves the error handling to distinguish between "no integ
manifest" and "failed to load manifest". For the former, we still go
down the legacy test route. For the latter we can now fail the test with
an accurate error message.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license