chore(cli): keep a legacy configuration copy (#702)
I am planning to make some API breaking changes to the CLI
`Configuration`. Because we are exporting this as a legacy export, we
want to keep the API stable. Therefore I am adding a hard copy of the
current state as a legacy export.

Also re-organizes all legacy code to make maintenance easier.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license