feat: modification is forbidden during a refactor (#667)
CloudFormation’s stack refactoring API does not allow modifications
(other than resource moves). Initially we wanted to give CDK users a
more flexible experience, by allowing them to move resources while also
adding and removing other resources (and, in some cases, updating as
well). These additions and removals were ignored by the toolkit, and
only what has moved was sent to the API for refactoring. As a result of
this flexibility, the refactor operation may leave the infrastructure in
a state that is not the same as the CDK application.

This becomes a problem when this hypothetical intermediate state has
cross-stack references that don’t exist in the CDK application. Due to
the way CloudFormation deals with references and exports, we would
either not be able to create this kind of reference, or create it in a
way that blocks subsequent deployments.

This PR changes the comparison algorithm, to enforce the same constraint
that the CloudFormation API does, but at the CDK level. Now the set of
deployed resources and the set of local resources should be the same.
Only resource moves are allowed.

The filtering strategy has also changed. Now, the stacks included in the
comparison are as follows:

**Local**: all local stacks, filtered according to the provided patterns
(if any)

**Deployed**: all deployed stacks whose name satisfy at least one of
these properties:
- there exists a local stack (after the filter) with the same name.
- was explicitly included by the user

Filters are provided to the CLI in the usual way, as a positional
parameter to the `refactor` command:

```shell
cdk refactor *Gamma
```

Additional stack names are provided via a new CLI option:

```shell
cdk refactor --additional-stack-name=Foo --additional-stack-name=Bar
```


Closes https://github.com/aws/aws-cdk-cli/issues/657.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>