refactor(cli): use IoHost in CdkToolkit (#703)
Replace old indirect logging calls with direct IoHost calls. 

Pulled out from https://github.com/aws/aws-cdk-cli/pull/695 for
review-ability.

This is adding an async `MockCloudExecutable.create()` and using it in
tests. The reason for this is that it will be needed for another part of
#695 and I don't want to revert the changes to the tests files in this
PR.

Includes a refactor of the file event batching test case for `watch` as
it was simply not working correctly before.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license