refactor(cli): use IoHost in CLI `Configuration` (#695)
Resolves #291 

Replace old indirect logging calls in `Configuration` with direct IoHost
calls. Because this was the lat place using the old logging helpers, we
can now remove them completely.

Creating a `Configuration` object is now async, therefore this change
requires the introduction of an async builder-method for
`MockCloudExecutable` and updates to any tests using it.

While the change to `Configuration` is breaking the existing API, we are
already exporting a hard-copy of the previous version for legacy users
(see #702).

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license